+++ impl/src/from_str.rs
@@ -5,10 +5,13 @@ use std::str::FromStr;
 use std::{collections::HashMap, iter};
 
 use proc_macro2::TokenStream;
-use quote::{quote, ToTokens};
-use syn::{parse_quote, spanned::Spanned as _};
+use quote::{format_ident, quote, ToTokens};
+use syn::{parse::Parse, parse_quote, spanned::Spanned as _};
 
-use crate::utils::Either;
+use crate::utils::{
+    attr::{self, ParseMultiple as _},
+    Either, GenericsSearch, Spanning,
+};
 
 /// Expands a [`FromStr`] derive macro.
 pub fn expand(input: &syn::DeriveInput, _: &'static str) -> syn::Result<TokenStream> {
@@ -35,6 +38,9 @@ struct ForwardExpansion<'i> {
 
     /// [`syn::Field`] representing the wrapped type to forward implementation on.
     inner: &'i syn::Field,
+
+    /// Optional [`attr::Error`] enabling conversion into a custom error type.
+    custom_error: Option<attr::Error>,
 }
 
 impl<'i> TryFrom<&'i syn::DeriveInput> for ForwardExpansion<'i> {
@@ -62,9 +68,14 @@ impl<'i> TryFrom<&'i syn::DeriveInput> for ForwardExpansion<'i> {
             ));
         };
 
+        let custom_error =
+            attr::Error::parse_attrs(&input.attrs, &format_ident!("from_str"))?
+                .map(Spanning::into_inner);
+
         Ok(Self {
             self_ty: (&input.ident, &input.generics),
             inner,
+            custom_error,
         })
     }
 }
@@ -75,8 +86,9 @@ impl ToTokens for ForwardExpansion<'_> {
         let inner_ty = &self.inner.ty;
         let ty = self.self_ty.0;
 
+        let generics_search = GenericsSearch::from(self.self_ty.1);
         let mut generics = self.self_ty.1.clone();
-        if !generics.params.is_empty() {
+        if generics_search.any_in(inner_ty) {
             generics.make_where_clause().predicates.push(parse_quote! {
                 #inner_ty: derive_more::core::str::FromStr
             });
@@ -85,14 +97,25 @@ impl ToTokens for ForwardExpansion<'_> {
 
         let constructor = self.inner.self_constructor([parse_quote! { v }]);
 
+        let mut error_ty =
+            quote! { <#inner_ty as derive_more::core::str::FromStr>::Err };
+        let mut error_conv = quote! {};
+        if let Some(custom_error) = &self.custom_error {
+            error_ty = custom_error.ty.to_token_stream();
+            error_conv = custom_error.conv.as_ref().map_or_else(
+                || quote! { .map_err(derive_more::core::convert::Into::into) },
+                |conv| quote! { .map_err(#conv) },
+            );
+        }
+
         quote! {
             #[automatically_derived]
             impl #impl_generics derive_more::core::str::FromStr for #ty #ty_generics #where_clause {
-                type Err = <#inner_ty as derive_more::core::str::FromStr>::Err;
+                type Err = #error_ty;
 
                 #[inline]
                 fn from_str(s: &str) -> derive_more::core::result::Result<Self, Self::Err> {
-                    derive_more::core::str::FromStr::from_str(s).map(|v| #constructor)
+                    derive_more::core::str::FromStr::from_str(s).map(|v| #constructor)#error_conv
                 }
             }
         }.to_tokens(tokens);
@@ -106,19 +129,34 @@ struct FlatExpansion<'i> {
     /// [`syn::Ident`]: struct@syn::Ident
     self_ty: (&'i syn::Ident, &'i syn::Generics),
 
-    /// [`syn::Ident`]s along with the matched values (enum variants or struct itself).
+    /// [`syn::Ident`]s along with the matched values (enum variants or struct itself), and
+    /// a value-specific [`attr::RenameAll`] overriding [`FlatExpansion::rename_all`], if any.
     ///
     /// [`syn::Ident`]: struct@syn::Ident
     matches: Vec<(
         &'i syn::Ident,
         Either<&'i syn::DataStruct, &'i syn::Variant>,
+        Option<attr::RenameAll>,
     )>,
+
+    /// [`FlatExpansion::matches`] grouped by its similar representation for detecting whether their
+    /// case-insensitivity should be disabled.
+    similar_matches: HashMap<String, Vec<&'i syn::Ident>>,
+
+    /// Optional [`attr::RenameAll`] indicating the case conversion to be applied to all the matched
+    /// values (enum variants or struct itself).
+    rename_all: Option<attr::RenameAll>,
+
+    /// Optional [`attr::Error`] enabling conversion into a custom error type.
+    custom_error: Option<attr::Error>,
 }
 
 impl<'i> TryFrom<&'i syn::DeriveInput> for FlatExpansion<'i> {
     type Error = syn::Error;
 
     fn try_from(input: &'i syn::DeriveInput) -> syn::Result<Self> {
+        let attr_ident = &format_ident!("from_str");
+
         let matches = match &input.data {
             syn::Data::Struct(data) => {
                 if !data.fields.is_empty() {
@@ -127,7 +165,7 @@ impl<'i> TryFrom<&'i syn::DeriveInput> for FlatExpansion<'i> {
                         "only structs with no fields can derive `FromStr`",
                     ));
                 }
-                vec![(&input.ident, Either::Left(data))]
+                vec![(&input.ident, Either::Left(data), None)]
             }
             syn::Data::Enum(data) => data
                 .variants
@@ -139,7 +177,10 @@ impl<'i> TryFrom<&'i syn::DeriveInput> for FlatExpansion<'i> {
                             "only enums with no fields can derive `FromStr`",
                         ));
                     }
-                    Ok((&variant.ident, Either::Right(variant)))
+                    let attr =
+                        attr::RenameAll::parse_attrs(&variant.attrs, attr_ident)?
+                            .map(Spanning::into_inner);
+                    Ok((&variant.ident, Either::Right(variant), attr))
                 })
                 .collect::<syn::Result<_>>()?,
             syn::Data::Union(_) => {
@@ -150,9 +191,69 @@ impl<'i> TryFrom<&'i syn::DeriveInput> for FlatExpansion<'i> {
             }
         };
 
+        let FlatContainerAttributes {
+            rename_all,
+            error: custom_error,
+        } = FlatContainerAttributes::parse_attrs(&input.attrs, attr_ident)?
+            .map(Spanning::into_inner)
+            .unwrap_or_default();
+
+        let mut similar_matches = <HashMap<_, Vec<_>>>::new();
+        if rename_all.is_none() {
+            for (ident, _, renaming) in &matches {
+                let name = ident.to_string();
+                let lowercased = name.to_lowercase();
+                if let Some(rename) = renaming {
+                    let renamed_lowercased = rename.convert_case(&name);
+                    if renamed_lowercased != lowercased {
+                        similar_matches
+                            .entry(renamed_lowercased)
+                            .or_default()
+                            .push(*ident);
+                    }
+                }
+                similar_matches.entry(lowercased).or_default().push(*ident);
+            }
+        }
+
+        let mut exact_matches = <HashMap<String, Vec<String>>>::new();
+        for (ident, _, renaming) in &matches {
+            let name = ident.to_string();
+            let exact = if let Some(default_renaming) = &rename_all {
+                renaming
+                    .as_ref()
+                    .unwrap_or(default_renaming)
+                    .convert_case(&name)
+            } else if let Some(renaming) = renaming {
+                renaming.convert_case(&name)
+            } else {
+                let lowercased = name.to_lowercase();
+                if similar_matches[&lowercased].len() > 1 {
+                    name.clone()
+                } else {
+                    lowercased
+                }
+            };
+            exact_matches.entry(exact).or_default().push(name);
+        }
+        if let Some((string, variants)) =
+            exact_matches.into_iter().find(|(_, vs)| vs.len() > 1)
+        {
+            return Err(syn::Error::new(
+                input.ident.span(),
+                format!(
+                    "`{}` variants cannot have the same \"{string}\" string representation",
+                    variants.join("`, `"),
+                ),
+            ));
+        }
+
         Ok(Self {
             self_ty: (&input.ident, &input.generics),
             matches,
+            similar_matches,
+            rename_all,
+            custom_error,
         })
     }
 }
@@ -165,40 +266,66 @@ impl ToTokens for FlatExpansion<'_> {
             self.self_ty.1.split_for_impl();
         let ty_name = ty.to_string();
 
-        let similar_lowercased = self
-            .matches
-            .iter()
-            .map(|(v, _)| v.to_string().to_lowercase())
-            .fold(<HashMap<_, u8>>::new(), |mut counts, v| {
-                *counts.entry(v).or_default() += 1;
-                counts
-            });
+        let scrutinee_lowercased = self
+            .rename_all
+            .is_none()
+            .then(|| quote! { .to_lowercase().as_str() });
+        let match_arms = if let Some(default_renaming) = self.rename_all {
+            self.matches
+                .iter()
+                .map(|(ident, value, renaming)| {
+                    let converted = renaming
+                        .unwrap_or(default_renaming)
+                        .convert_case(&ident.to_string());
+                    let constructor = value.self_constructor_empty();
 
-        let match_arms = self.matches.iter().map(|(ident, value)| {
-            let name = ident.to_string();
-            let lowercased = name.to_lowercase();
+                    quote! { #converted => #constructor, }
+                })
+                .collect::<Vec<_>>()
+        } else {
+            self.matches
+                .iter()
+                .map(|(ident, value, renaming)| {
+                    let name = ident.to_string();
+                    let constructor = value.self_constructor_empty();
+                    if let Some(rename) = renaming {
+                        let exact_name = rename.convert_case(&name);
 
-            let exact_guard =
-                (similar_lowercased[&lowercased] > 1).then(|| quote! { if s == #name });
-            let constructor = value.self_constructor_empty();
+                        quote! { _ if s == #exact_name => #constructor, }
+                    } else {
+                        let lowercased = name.to_lowercase();
+                        let exact_guard = (self.similar_matches[&lowercased].len() > 1)
+                            .then(|| quote! { if s == #name });
 
-            quote! { #lowercased #exact_guard => #constructor, }
-        });
+                        quote! { #lowercased #exact_guard => #constructor, }
+                    }
+                })
+                .collect()
+        };
+
+        let default_error = quote! { derive_more::FromStrError::new(#ty_name) };
+        let mut error_ty = quote! { derive_more::FromStrError };
+        let mut error_val = default_error.clone();
+        if let Some(custom_error) = &self.custom_error {
+            error_ty = custom_error.ty.to_token_stream();
+            error_val = custom_error.conv.as_ref().map_or_else(
+                || quote! { derive_more::core::convert::Into::into(#default_error) },
+                |conv| quote! { (#conv)(#default_error) },
+            );
+        }
 
         quote! {
             #[allow(unreachable_code)] // for empty enums
             #[automatically_derived]
             impl #impl_generics derive_more::core::str::FromStr for #ty #ty_generics #where_clause {
-                type Err = derive_more::FromStrError;
+                type Err = #error_ty;
 
-                fn from_str(
-                    s: &str,
-                ) -> derive_more::core::result::Result<Self, derive_more::FromStrError> {
-                    derive_more::core::result::Result::Ok(match s.to_lowercase().as_str() {
+                fn from_str(s: &str) -> derive_more::core::result::Result<
+                    Self, <Self as derive_more::core::str::FromStr>::Err,
+                > {
+                    derive_more::core::result::Result::Ok(match s #scrutinee_lowercased {
                         #( #match_arms )*
-                        _ => return derive_more::core::result::Result::Err(
-                            derive_more::FromStrError::new(#ty_name),
-                        ),
+                        _ => return derive_more::core::result::Result::Err(#error_val),
                     })
                 }
             }
@@ -335,5 +462,95 @@ impl<L: FieldsExt, R: FieldsExt> FieldsExt for Either<&L, &R> {
             Self::Left(l) => l.self_ty(),
             Self::Right(r) => r.self_ty(),
         }
+    }
+}
+
+/// Representation of possible [`FromStr`] derive macro attributes placed on an enum or a struct for
+/// a [`FlatExpansion`].
+///
+/// ```rust,ignore
+/// #[<attribute>(rename_all = "<casing>")]
+/// #[<attribute>(error(<ty>))]
+/// #[<attribute>(error(<ty>, <conv>))]
+/// ```
+///
+/// Both `#[<attribute>(rename_all = "<casing>")]` and `#[<attribute>(error(<ty>[, <conv>]))]` can
+/// be specified only once.
+#[derive(Default)]
+struct FlatContainerAttributes {
+    /// [`attr::RenameAll`] for case conversion.
+    rename_all: Option<attr::RenameAll>,
+
+    /// [`attr::Error`] for conversion into a custom error type.
+    error: Option<attr::Error>,
+}
+
+impl Parse for FlatContainerAttributes {
+    fn parse(input: syn::parse::ParseStream) -> syn::Result<Self> {
+        mod ident {
+            use syn::custom_keyword;
+
+            custom_keyword!(error);
+            custom_keyword!(rename_all);
+        }
+
+        let ahead = input.lookahead1();
+        if ahead.peek(ident::error) {
+            Ok(Self {
+                error: Some(input.parse()?),
+                ..Default::default()
+            })
+        } else if ahead.peek(ident::rename_all) {
+            Ok(Self {
+                rename_all: Some(input.parse()?),
+                ..Default::default()
+            })
+        } else {
+            Err(ahead.error())
+        }
+    }
+}
+
+impl attr::ParseMultiple for FlatContainerAttributes {
+    fn merge_attrs(
+        prev: Spanning<Self>,
+        new: Spanning<Self>,
+        name: &syn::Ident,
+    ) -> syn::Result<Spanning<Self>> {
+        let Spanning {
+            span: prev_span,
+            item: mut prev,
+        } = prev;
+        let Spanning {
+            span: new_span,
+            item: new,
+        } = new;
+
+        if new
+            .rename_all
+            .and_then(|n| prev.rename_all.replace(n))
+            .is_some()
+        {
+            return Err(syn::Error::new(
+                new_span,
+                format!("multiple `#[{name}(rename_all=\"...\")]` attributes aren't allowed"),
+            ));
+        }
+
+        if prev.error.is_some() && new.error.is_some() {
+            return Err(syn::Error::new(
+                new_span,
+                format!(
+                    "multiple `#[{name}(error(\"...\")]` attributes aren't allowed",
+                ),
+            ));
+        }
+
+        prev.error = prev.error.or(new.error);
+
+        Ok(Spanning::new(
+            prev,
+            prev_span.join(new_span).unwrap_or(prev_span),
+        ))
     }
 }