+++ CHANGELOG.md
@@ -12,10 +12,27 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Support `#[display(rename_all = "<casing>")]` attribute to change output for
   implicit naming of unit enum variants or unit structs when deriving `Display`.
   ([#443](https://github.com/JelteF/derive_more/pull/443))
+- Support `#[from_str(rename_all = "<casing>")]` attribute for unit enum variants
+  and unit structs when deriving `FromStr`.
+  ([#467](https://github.com/JelteF/derive_more/pull/467))
 - Support `Option` fields for `Error::source()` in `Error` derive.
   ([#459](https://github.com/JelteF/derive_more/pull/459))
 - Support structs with no fields in `FromStr` derive.
   ([#469](https://github.com/JelteF/derive_more/pull/469))
+- Add `PartialEq` derive similar to `std`'s one, but considering generics correctly,
+  and implementing `ne()` method as well.
+  ([#473](https://github.com/JelteF/derive_more/pull/473),
+   [#475](https://github.com/JelteF/derive_more/pull/475))
+- Add `Eq` derive similar to `std`'s one, but considering generics correctly.
+  ([#479](https://github.com/JelteF/derive_more/pull/479))
+- Proxy-pass `#[allow]`/`#[expect]` attributes of the type in `Constructor` derive.
+  ([#477](https://github.com/JelteF/derive_more/pull/477))
+- Support `Deref` and `DerefMut` derives for enums.
+  ([#485](https://github.com/JelteF/derive_more/pull/485))
+- Support custom error in `FromStr` derive.
+  ([#494](https://github.com/JelteF/derive_more/pull/494))
+- Support custom error in `TryInto` derive.
+  ([#503](https://github.com/JelteF/derive_more/pull/503))
 - Support `#[skip]` attribute for the `Add` derive to allow skipping zero sized
   fields. ([#472](https://github.com/JelteF/derive_more/pull/472))
 
@@ -30,6 +47,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
   ([#454](https://github.com/JelteF/derive_more/pull/454))
 - Silent no-op when `#[try_from(repr)]` attribute is not specified for `TryFrom` derive.
   ([#458](https://github.com/JelteF/derive_more/pull/458))
+- Missing trait bounds in `AsRef`/`AsMut` derives when associative types are involved.
+  ([#474](https://github.com/JelteF/derive_more/pull/474))
+- Erroneous code generated in `Try`/`TryInto` derives when `Self` type is present in
+  the struct or enum definition.
+  ([#489](https://github.com/JelteF/derive_more/pull/489))
+- Dependency on unstable `feature(error_generic_member_access)` in `Error` derive when
+  using `Backtrace` on a non-nightly toolchain.
+  ([#513](https://github.com/JelteF/derive_more/pull/513))
 
 ## 2.0.1 - 2025-02-03